+++ src/main/java/com/terminalvelocitycabbage/engine/client/renderer/shapes/Rectangle.java
@@ -10,7 +10,7 @@ public class Rectangle {
 		return new Model.Part(
 			new MeshPart(
 				new Vertex[]{ topLeft, bottomLeft, bottomRight, topRight },
-				new short[] { 0, 1, 2, 2, 3, 0 }
+				new int[] { 0, 1, 2, 2, 3, 0 }
 			)
 		);
 	}
